pyspacer training: read parquet annotations in chunks with DuckDB, to rein in memory usage during that step.
Memory usage of the labels structure going into pyspacer is another story, but will address that another time.

Generators are now used to encapsulate the chunking, as well as the detection of when one image's annotations are done and another one starts. The latter helped fix a bug where a skipped feature vector would still contribute to the annotation counts.