Annotation viewing: various plot changes/reworks.
Abandon the idea of box markers that are to-scale with the feature patches, since it's too hard to get that right.

Use shape to indicate the score range of the annotation.

Use edge properties and path effects for outlines, rather than hackily drawing 'shadow versions' of the markers/text.

Move a lot of plotting specifics from pyspacer/annotation.py to common/plots.py.

Don't show axis ticks/labels.