More DuckDB and less in-Python manipulation for the code that detects/handles missing feature vectors.
Also add tests for handle_missing_feature_vectors(), since it's tricky to manually test that functionality. These happen to be our first unit tests. The main shortcoming with them is that they create a train.log file at the cwd, with no way as of yet to configure it to go elsewhere.